+++ Source/Utils/IOptionalPlugin.cs
@@ -1,11 +1,9 @@
-using ImGuiNET;
-
 namespace UImGui
 {
 	internal interface IOptionalPlugin
 	{
-		void Create(Context context);
-		void SetCurrent(Context context);
-		void Destroy(Context context);
+		void CreateContext(Context context);
+		void SetCurrentContext(Context context);
+		void DestroyContext(Context context);
 	}
 }